Extract editor main current file view
